requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
streamlit>=1.30.0
gspread>=6.0.0
google-auth>=2.0.0
